+++ server/src/test/java/org/opensearch/common/xcontent/JsonToStringXContentParserTests.java
@@ -12,13 +12,15 @@ import org.opensearch.common.xcontent.json.JsonXContent;
 import org.opensearch.core.xcontent.DeprecationHandler;
 import org.opensearch.core.xcontent.XContentBuilder;
 import org.opensearch.core.xcontent.XContentParser;
+import org.opensearch.index.mapper.MapperParsingException;
 import org.opensearch.test.OpenSearchTestCase;
+import org.hamcrest.Matchers;
 
 import java.io.IOException;
 
 public class JsonToStringXContentParserTests extends OpenSearchTestCase {
 
-    private String flattenJsonString(String fieldName, String in) throws IOException {
+    private String flattenJsonString(String fieldName, String in, int depthLimit, String nullValue, int ignoreAbove) throws IOException {
         try (
             XContentParser parser = JsonXContent.jsonXContent.createParser(
                 xContentRegistry(),
@@ -30,7 +32,9 @@ public class JsonToStringXContentParserTests extends OpenSearchTestCase {
                 xContentRegistry(),
                 DeprecationHandler.THROW_UNSUPPORTED_OPERATION,
                 parser,
-                fieldName
+                fieldName,
+                depthLimit,
+                ignoreAbove
             );
             // Point to the first token (should be START_OBJECT)
             jsonToStringXContentParser.nextToken();
@@ -53,7 +57,7 @@ public class JsonToStringXContentParserTests extends OpenSearchTestCase {
                 + "\"flat._value\":[\"1\",\"2.0\",\"three\"],"
                 + "\"flat._valueAndPath\":[\"flat.second.inner=2.0\",\"flat.first=1\",\"flat.third=three\"]"
                 + "}",
-            flattenJsonString("flat", jsonExample)
+            flattenJsonString("flat", jsonExample, 5, null, 100)
         );
     }
 
@@ -68,7 +72,7 @@ public class JsonToStringXContentParserTests extends OpenSearchTestCase {
                 + "\"flat._value\":[\"1\",\"2.0\",\"three\"],"
                 + "\"flat._valueAndPath\":[\"flat.second.inner=2.0\",\"flat.first=1\",\"flat.third=three\"]"
                 + "}",
-            flattenJsonString("flat", jsonExample)
+            flattenJsonString("flat", jsonExample, 5, null, 100)
         );
     }
 
@@ -87,7 +91,7 @@ public class JsonToStringXContentParserTests extends OpenSearchTestCase {
                 + "\"flat._value\":[\"1\",\"2.0\",\"three\"],"
                 + "\"flat._valueAndPath\":[\"flat.first=1\",\"flat.second.inner.really_inner=2.0\",\"flat.third=three\"]"
                 + "}",
-            flattenJsonString("flat", jsonExample)
+            flattenJsonString("flat", jsonExample, 5, null, 100)
         );
     }
 
@@ -106,7 +110,7 @@ public class JsonToStringXContentParserTests extends OpenSearchTestCase {
                 + "\"flat._value\":[\"1\",\"2.0\",\"three\"],"
                 + "\"flat._valueAndPath\":[\"flat.first=1\",\"flat.second.inner.totally.absolutely.inner=2.0\",\"flat.third=three\"]"
                 + "}",
-            flattenJsonString("flat", jsonExample)
+            flattenJsonString("flat", jsonExample, 5, null, 100)
         );
     }
 
@@ -128,7 +132,7 @@ public class JsonToStringXContentParserTests extends OpenSearchTestCase {
                 + "\"flat._valueAndPath\":["
                 + "\"flat.field.detail.foooooooooooo.name=baz\""
                 + "]}",
-            flattenJsonString("flat", jsonExample)
+            flattenJsonString("flat", jsonExample, 5, null, 100)
         );
     }
 
@@ -158,7 +162,58 @@ public class JsonToStringXContentParserTests extends OpenSearchTestCase {
                 + "\"flat.field.numbers=3\","
                 + "\"flat.field.numbers=2\""
                 + "]}",
-            flattenJsonString("flat", jsonExample)
+            flattenJsonString("flat", jsonExample, 5, null, 100)
+        );
+    }
+
+    public void testDepthLimit() throws IOException {
+        String jsonExample = "{"
+            + "\"first\" : \"1\","
+            + "\"second.inner\" : {"
+            + "  \"totally.absolutely.inner\" : \"2.0\""
+            + "},"
+            + "\"third\": \"three\""
+            + "}";
+        MapperParsingException e = expectThrows(MapperParsingException.class, () -> flattenJsonString("flat", jsonExample, 1, null, 100));
+        assertThat(
+            e.getRootCause().getMessage(),
+            Matchers.containsString("the depth of flat_object field path [flat, second.inner] is bigger than maximum depth [1]")
+        );
+        assertEquals(
+            "{"
+                + "\"flat\":[\"third\",\"absolutely\",\"totally\",\"inner\",\"first\",\"second\"],"
+                + "\"flat._value\":[\"1\",\"2.0\",\"three\"],"
+                + "\"flat._valueAndPath\":[\"flat.first=1\",\"flat.second.inner.totally.absolutely.inner=2.0\",\"flat.third=three\"]"
+                + "}",
+            flattenJsonString("flat", jsonExample, 3, null, 100)
+        );
+    }
+
+    public void testIgnoreAbove() throws IOException {
+        String jsonExample = "{"
+            + "\"first\" : \"1\","
+            + "\"second.inner\" : {"
+            + "  \"totally.absolutely.inner\" : \"2.0\""
+            + "},"
+            + "\"third\": \"three\""
+            + "}";
+
+        assertEquals(
+            "{"
+                + "\"flat\":[\"third\",\"absolutely\",\"totally\",\"inner\",\"first\",\"second\"],"
+                + "\"flat._value\":[\"1\",\"2.0\",\"three\"],"
+                + "\"flat._valueAndPath\":[\"flat.first=1\",\"flat.second.inner.totally.absolutely.inner=2.0\",\"flat.third=three\"]"
+                + "}",
+            flattenJsonString("flat", jsonExample, 5, null, 5)
+        );
+
+        assertEquals(
+            "{"
+                + "\"flat\":[\"absolutely\",\"totally\",\"inner\",\"first\",\"second\"],"
+                + "\"flat._value\":[\"1\",\"2.0\"],"
+                + "\"flat._valueAndPath\":[\"flat.first=1\",\"flat.second.inner.totally.absolutely.inner=2.0\"]"
+                + "}",
+            flattenJsonString("flat", jsonExample, 5, null, 4)
         );
     }
 }